+++ Locale/MRP_Locale.zhTW.lua
@@ -1,5 +1,10 @@
 -- MRP_Locale.zhTW.lua
-local _, MRP = ...
+-- local _, MRP = ...
+
+--[[
+  Contributors:
+  * fang2hou (https://github.com/fang2hou)
+]]
 
 if (GetLocale() ~= "zhTW") then
   return;
@@ -125,7 +130,6 @@ MRP.L = {
   ["Waypoint_1003"] = "從%s搭船到%s",
   ["Waypoint_1004"] = "從%s搭飛船到%s",
   ["Waypoint_1005"] = "使用%s到%s",
-  ["Waypoint_1006"] = "施放%s到%s",
   ["Waypoint_Instance"] = "進入%s",
   ["Waypoint_WorldBoss"] = "擊殺%s",
   ["Wrong Difficulty"] = "錯誤難度",